added confirm() to delete since it's a permanent operation; sort user list on email; corrected refresh to hide the createuser panel;

+++ app/features/users/users-ctrl.js
@@ -27,6 +27,7 @@ app.controller('UsersCtrl', function($scope, $u, UsersFactory) {
   };
 
   $scope.delete = function(user) {
+    if(!confirm('Are you sure you want to delete '+user.email+'?')) { return; }
     UsersFactory.deleteUser(user.email,
       function() { // success
         $u.remove($scope.users, { email: user.email });
